ai_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
requests==2.32.3
openai-whisper==20231117
pydantic==2.9.2
